Refactored dags into the example_dags folder.
