+++ google/gmp/example_dags/dv360_create_sdf_advertisers_report_dag.py
@@ -16,8 +16,6 @@
 #
 ###########################################################################
 """Example DAG which creates DV360 report."""
-import sys
-sys.path.append("..")
 from datetime import datetime
 from datetime import timedelta
 from airflow import DAG
