+++ google/gmp/example_dags/dv360_get_sdf_advertisers_from_report_dag.py
@@ -16,8 +16,6 @@
 #
 ###########################################################################
 """Example DAG which extracts Advertiser IDs from a DV360 report."""
-import sys
-sys.path.append('..')
 from datetime import datetime
 from datetime import timedelta
 from airflow import DAG
